Разделил API на разные классы
