+++ app/src/main/java/org/hse/ataskmobileclient/apis/TasksApi.kt
@@ -14,20 +14,7 @@ import org.hse.ataskmobileclient.models.AuthResult
 import org.hse.ataskmobileclient.models.TaskResult
 import org.hse.ataskmobileclient.models.Urls
 
-class API {
-
-    fun authWithGoogle (token: String, responseHandler: (result: AuthResult) -> Unit?) {
-
-        val authDto = AuthDto(token)
-
-        Urls()
-            .getGoogleAuthUrl()
-            .httpPost()
-            .jsonBody(authDto)
-            .responseString { _, _, result ->
-                this.authResultHandler(result, responseHandler)
-            }
-    }
+class TasksApi {
 
     fun createTask(
         token: String,
@@ -60,32 +47,13 @@ class API {
         )
     }
 
-    private fun authResultHandler (
-        result : Result<String, FuelError>,
-        responseHandler: (result: AuthResult) -> Unit?
-    ) {
-        when (result) {
-            is Result.Failure -> {
-                Log.i("ErrorMsg", result.getException().message)
-                result.getException().stackTrace
-                throw Exception(result.getException())
-            }
-            is Result.Success -> {
-                val authResult: AuthResult? = Gson().fromJson(result.get(), AuthResult::class.java)
-                if (authResult != null) {
-                    responseHandler.invoke(authResult)
-                }
-            }
-        }
-    }
-
     private fun tasksResultHandler (
         result : Result<String, FuelError>,
         responseHandler: (result: ArrayList<TaskResult>) -> Unit?
     )  {
         when (result) {
             is Result.Failure -> {
-                Log.i("ErrorMsg", result.getException().message)
+                Log.i("ErrorMsg", result.getException().message ?: "")
                 result.getException().stackTrace
                 throw Exception(result.getException())
             }
@@ -98,6 +66,6 @@ class API {
     }
 
     companion object {
-        private const val TAG = "API"
+        private const val TAG = "TasksApi"
     }
 }
